Testing 1
Added enzyme tests and configuration.  Tests create a screenshot of the render jsx and compare to make sure there werent and changes.  Added cardlist test that mocks the robots array props and tests alternative loading state.  Added mock function that tests the updateCount method in the counterButton component to make sure the count is updated.